Fix some typos
I didn't change this one, but I'd like to ask any reviewer to consider the change. This expression was confusing Spanish speaker for Spain: 'Ten en cuenta que estas funciones son nuestra cerca', since we don't use 'cerca' as fence, but only as close (not far). I realised that it's also used at the beginning of the chapter, but that time I read 'cera' (wax), which is probably a good translation since it carries the same message, being the most basic set of tools that can be used.